+++ src/main/java/io/github/adex720/minigames/gameplay/profile/stat/StatList.java
@@ -19,8 +19,8 @@ public class StatList {
 
     private final MinigamesBot bot;
 
-    private final HashMap<String, Value<Integer>> statsByName;
-    private final HashMap<Integer, Value<Integer>> statsById;
+    private final HashMap<String, Value<Long>> statsByName;
+    private final HashMap<Integer, Value<Long>> statsById;
 
     public StatList(MinigamesBot bot) {
         this.bot = bot;
@@ -29,7 +29,7 @@ public class StatList {
         statsById = new HashMap<>();
 
         for (Stat stat : bot.getStatManager().getAll()) {
-            Value<Integer> value = new Value<>(0);
+            Value<Long> value = new Value<>(0L);
 
             statsByName.put(stat.name(), value);
             statsById.put(stat.id(), value);
@@ -43,7 +43,7 @@ public class StatList {
         statsById = new HashMap<>();
 
         for (Stat stat : bot.getStatManager().getAll()) {
-            Value<Integer> value = new Value<>(JsonHelper.getInt(json, Integer.toString(stat.id()), 0));
+            Value<Long> value = new Value<>(JsonHelper.getLong(json, Long.toString(stat.id()), 0));
 
             statsByName.put(stat.name(), value);
             statsById.put(stat.id(), value);
@@ -51,18 +51,18 @@ public class StatList {
     }
 
     @CheckReturnValue
-    public int getValue(String stat) {
+    public long getValue(String stat) {
         return statsByName.get(stat).value;
     }
 
     @CheckReturnValue
-    public int getValue(int stat) {
+    public long getValue(int stat) {
         return statsById.get(stat).value;
     }
 
-    public int increaseStat(Stat stat, Profile profile) {
+    public long increaseStat(Stat stat, Profile profile) {
         statsByName.get(stat.name()).value++;
-        int newValue = statsById.get(stat.id()).value++;
+        long newValue = statsById.get(stat.id()).value++;
 
         if (stat.onLeaderboard()) {
             bot.getStatManager().getLeaderboard(stat.id()).update(profile);
@@ -70,9 +70,9 @@ public class StatList {
         return newValue;
     }
 
-    public int increaseStat(Stat stat, int amount, Profile profile) {
+    public long increaseStat(Stat stat, long amount, Profile profile) {
         statsByName.get(stat.name()).value += amount;
-        int newValue = statsById.get(stat.id()).value += amount;
+        long newValue = statsById.get(stat.id()).value += amount;
 
         if (stat.onLeaderboard()) {
             bot.getStatManager().getLeaderboard(stat.id()).update(profile);
@@ -80,23 +80,23 @@ public class StatList {
         return newValue;
     }
 
-    public int increaseStat(String stat, Profile profile) {
+    public long increaseStat(String stat, Profile profile) {
         return increaseStat(bot.getStatManager().get(stat), profile);
     }
 
-    public int increaseStat(String stat, int amount, Profile profile) {
+    public long increaseStat(String stat, long amount, Profile profile) {
         return increaseStat(bot.getStatManager().get(stat), amount, profile);
     }
 
-    public int increaseStat(int stat, Profile profile) {
+    public long increaseStat(int stat, Profile profile) {
         return increaseStat(bot.getStatManager().get(stat), profile);
     }
 
-    public int increaseStat(int stat, int amount, Profile profile) {
+    public long increaseStat(int stat, long amount, Profile profile) {
         return increaseStat(bot.getStatManager().get(stat), amount, profile);
     }
 
-    public void setValue(Stat stat, int value, final Profile profile) {
+    public void setValue(Stat stat, long value, final Profile profile) {
         statsByName.get(stat.name()).value = value;
         statsById.get(stat.id()).value = value;
 
@@ -105,27 +105,27 @@ public class StatList {
         }
     }
 
-    public void setValue(int stat, int value, final Profile profile) {
+    public void setValue(int stat, long value, final Profile profile) {
         setValue(bot.getStatManager().get(stat), value, profile);
     }
 
-    public void setValue(String stat, int value, final Profile profile) {
+    public void setValue(String stat, long value, final Profile profile) {
         setValue(bot.getStatManager().get(stat), value, profile);
     }
 
     public JsonObject asJson() {
         JsonObject json = new JsonObject();
 
-        for (Map.Entry<Integer, Value<Integer>> entry : statsById.entrySet()) {
-            int value = entry.getValue().value;
+        for (Map.Entry<Integer, Value<Long>> entry : statsById.entrySet()) {
+            long value = entry.getValue().value;
             if (value == 0) continue;
-            json.addProperty(Integer.toString(entry.getKey()), value);
+            json.addProperty(Long.toString(entry.getKey()), value);
         }
 
         return json;
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return statsById.size();
     }
 }
